requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
PyMySQL==1.1.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
requests==2.31.0